code/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
aiohttp==3.10.10
python-dotenv==1.0.1
